+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - More PvP achievements (6)
 - Pet Battle achievements (1)
 - Misc achievements (7)
+- Option to automatically adjust the summary achievements list to fit the space without showing a scroll bar
 
 ### Changed
 - Tons of back end changes